+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Changed
+- Ability to use weighted random strategy in settings
+
 ## Fixed
 - Increased performance for function that removes markdown in author's name
+
 
 ## [1.3.4] - 2022/06/20
 ## Changed